refactor(ns04): consolidate closure packet
