+++ tests/test_session_runner_pending_permissions.py
@@ -1,7 +1,8 @@
 from __future__ import annotations
-import asyncio, multiprocessing, os, stat, threading, pytest; from pathlib import Path; from breadboard.product.harness.lock import EffectiveHarnessLock; from breadboard.product.runtime.session import Session as ProductSession
+import asyncio, json, multiprocessing, os, stat, threading, pytest; from pathlib import Path; from breadboard.product.harness.lock import EffectiveHarnessLock; from breadboard.product.runtime.events import Session as ProductSession
 from agentic_coder_prototype.api.cli_bridge.events import EventType; from agentic_coder_prototype.permissions import load_permission_rules, upsert_permission_rule; from agentic_coder_prototype.permissions import rules_store; from agentic_coder_prototype.permissions.broker import PermissionBroker; from agentic_coder_prototype.permissions.rules_store import RULES_REL_PATH, _locked_rules
 from agentic_coder_prototype.api.cli_bridge.models import SessionCreateRequest, SessionStatus; from agentic_coder_prototype.api.cli_bridge.registry import SessionRecord, SessionRegistry; from agentic_coder_prototype.api.cli_bridge.session_runner import SessionRunner, _canonical_permission_resolution
+from agentic_coder_prototype.api.cli_bridge.service import SessionService
 def _runner(session_id: str = "session") -> SessionRunner: return SessionRunner(session=SessionRecord(session_id=session_id, status=SessionStatus.RUNNING), registry=SessionRegistry(), request=SessionCreateRequest(config_path="dummy.yml", task="task", stream=False))
 def _product_runner(session_id: str) -> tuple[SessionRunner, ProductSession]: runner = _runner(session_id); session = ProductSession.start(EffectiveHarnessLock._from_record({"graph_hash": "sha256:" + "a" * 64}), "task", session_id=session_id); runner.session.product_session = session; return runner, session
 async def _initialized() -> None: pass
@@ -127,3 +128,21 @@ def test_runner_serializes_concurrent_product_transitions() -> None:
         except BaseException as error: errors.append(error)
     threads = [threading.Thread(target=submit, args=(content,), daemon=True) for content in ("first", "second")]; [thread.start() for thread in threads]; [thread.join(timeout=2) for thread in threads]
     assert gate.attempts == 2 and not errors and all(not thread.is_alive() for thread in threads); assert ([event.sequence for event in session.events], runner._input_queue.qsize()) == ([1, 2, 3], 2)
+@pytest.mark.asyncio
+async def test_mutating_loader_cannot_split_frozen_start_artifacts(monkeypatch, tmp_path: Path) -> None:
+    first = {"tools": {"registry": {"include": ["first_tool"]}}, "policies": {"tools": {"allow": ["first_tool"], "deny": ["blocked_tool"]}, "models": {"deny": ["blocked-model"]}, "approvals": {"mode": "always_required"}}, "provider_auth_runtime": {"openai": {"api_key": "runtime-secret"}}, "provider_auth_runtime.openai.api_key": "flat-secret", "wrapper": {"provider_auth_runtime": {"openai": {"api_key": "nested-secret"}}}}
+    source, calls = json.loads(json.dumps(first)), []
+    def load(path): calls.append(path); return source  # type: ignore[no-untyped-def]
+    def enable_primitives() -> bool: source["tools"]["registry"]["include"] = ["second_tool"]; source["policies"]["tools"] = {"allow": ["second_tool"], "deny": []}; source["policies"]["models"] = {"deny": []}; return True
+    async def start(_runner) -> None: return None  # type: ignore[no-untyped-def]
+    monkeypatch.setattr("agentic_coder_prototype.api.cli_bridge.session_runner.load_agent_config", load); monkeypatch.setattr("agentic_coder_prototype.api.cli_bridge.runtime_emission.load_agent_config", lambda _: pytest.fail("emitter reloaded mutable source")); monkeypatch.setattr("agentic_coder_prototype.api.cli_bridge.service.primitive_emission_enabled", enable_primitives)
+    monkeypatch.setattr("agentic_coder_prototype.api.cli_bridge.service.uuid.uuid4", lambda: "frozen-session"); monkeypatch.setattr("agentic_coder_prototype.api.cli_bridge.session_runner.SessionRunner.start", start); monkeypatch.setenv("BREADBOARD_RUNTIME_RECORD_ROOT", str(tmp_path / "records")); monkeypatch.setenv("BREADBOARD_SESSION_EVENT_ROOT", str(tmp_path / "events"))
+    service = SessionService(); response = await service.create_session(SessionCreateRequest(config_path="mutating.json", task="test")); record = await service.ensure_session(response.session_id); root = tmp_path / "records" / response.session_id; names = ("effective_config_graph", "capability_registry", "effective_tool_surface", "effective_operation_policy"); payloads = {name: json.loads((root / f"{name}.json").read_text()) for name in names}
+    graph, policy = payloads["effective_config_graph"], payloads["effective_operation_policy"]; tool_ids = {item["capability_id"] for item in payloads["capability_registry"]["capabilities"] if item["capability_type"] == "tool"}; assert calls == ["mutating.json"] and graph["graph_hash"] == record.product_session.events[0].payload["effective_lock_hash"]
+    assert tool_ids == {"tool.first_tool"} == set(payloads["effective_tool_surface"]["tool_ids"]) and [(rule["decision"], rule["match"]["pattern"]) for rule in policy["tool_policy"]["rules"]] == [("allow", "first_tool"), ("deny", "blocked_tool")]
+    serialized = json.dumps(payloads) + "".join(path.read_text() for path in tmp_path.rglob("*") if path.is_file()); assert policy["approvals"]["mode"] == "always_required" and all(secret not in serialized for secret in ("provider_auth_runtime", "runtime-secret", "flat-secret", "nested-secret", "second_tool"))
+    captured = {}; Agent = type("Agent", (), {"workspace_dir": str(tmp_path / "workspace"), "initialize": lambda self: None}); factory = lambda path, _workspace, _overrides: (captured.update(config=json.loads(Path(path).read_text()), path=path), Agent())[1]  # type: ignore[assignment]
+    record.runner.agent_factory = factory; record.runner.get_skill_catalog(); await record.runner._ensure_agent_initialized()
+    persisted = "".join(path.read_text() for root in (tmp_path / "records", tmp_path / "events") for path in root.rglob("*") if path.is_file()); assert calls == ["mutating.json"] and captured["config"] == record.runner.current_runtime_config() and captured["config"]["provider_auth_runtime"]["openai"]["api_key"] == "runtime-secret" and all(secret not in persisted for secret in ("provider_auth_runtime", "runtime-secret", "flat-secret", "nested-secret")) and not Path(captured["path"]).exists()
+    with pytest.raises(ValueError, match="denied by policy"): await record.runner.handle_command("set_model", {"model": "blocked-model"})
+    await service.stop_session(response.session_id); await record.event_queue.put(None); await record.dispatcher_task
